Rename all instances of 'scrapp*'
